% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pop.R
\name{generate_pop}
\alias{generate_pop}
\title{Simulate a baseline population}
\usage{
generate_pop(params, pop_size = NULL)
}
\arguments{
\item{params}{a parameters object, read in from a yaml file using \code{read_yaml()}}

\item{pop_size}{numeric - leave blank to use the value in the params file}
}
\value{
a dataframe
}
\description{
Simulate a baseline population
}
\examples{
params <- read_yaml('params.yaml')
generate_pop(params)
}
