% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{sample_segment_age}
\alias{sample_segment_age}
\title{Title sample a vector of segments using age groups}
\usage{
sample_segment_age(age, params, n)
}
\arguments{
\item{params}{parameters object}

\item{vector}{of age groups}
}
\value{
a vector of segments
}
\description{
Title sample a vector of segments using age groups
}
\examples{
sample_segment_age(pop$age, params)
}
