% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{sample_segment_age_wl}
\alias{sample_segment_age_wl}
\title{Title sample a vector of waiting list booleans using age and segment}
\usage{
sample_segment_age_wl(age, segment, params)
}
\arguments{
\item{age}{vector of age groups}

\item{segment}{vector of segment states}

\item{params}{parameters object}
}
\value{
a vector of booleans
}
\description{
Title sample a vector of waiting list booleans using age and segment
}
\examples{
sample_segment_age_wl(pop$age,pop$segment,params)
}
