% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_day.R
\name{sim_day}
\alias{sim_day}
\title{Simulate a day}
\usage{
sim_day(pop, day, params)
}
\arguments{
\item{pop}{df generated by \code{generate_pop}}

\item{day}{a day of the week, lowercase as per params file}

\item{params}{params file}
}
\value{
a data frame with outcomes appended
}
\description{
Simulate a day
}
\examples{
sim_day(pop,'wednesday',params)
}
