% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_period.R
\name{sim_period}
\alias{sim_period}
\title{Title simulate a period of time}
\usage{
sim_period(pop, ndays, params)
}
\arguments{
\item{pop}{df generated using \code{generate_pop()}}

\item{ndays}{number of days to simulate}

\item{params}{params file}
}
\value{
stacked data frame per patient per day
}
\description{
Title simulate a period of time
}
\examples{
pop <- generate_pop(params)
sim_period(pop,7,params)

}
